yolo-cpu/requirements.txt
ultralytics
wandb
torch
torchvision
opencv-python
Pillow
numpy
matplotlib
pyyaml
python-dotenv
ray[default]